Move test code into header and template it
